Ensure empty packages do not cause failure
refs: https://github.com/TryGhost/DevOps/issues/56

When we restore cache using restore-keys, we'll sometimes pull dependencies for packages which aren't in the current branch. This is fine, since we still install the correct packages, but it should be faster than starting from scratch.

Unfortunately, this sometimes means we'll have package directories that only contain a node_modules folder, with no package.json or code. They won't ever be used, but they were causing the call to monobundle to error out.

This change allows us to keep the speed of using the cache and correctly avoid trying to bundle these non-packages.